+++ eds-user-20210308/alibabacloud_eds_user20210308/models.py
@@ -2951,6 +2951,7 @@ class DescribeOrgsRequest(TeaModel):
         next_token: str = None,
         org_name: str = None,
         parent_org_id: str = None,
+        show_extras: Dict[str, Any] = None,
     ):
         # The maximum number of entries to return. Valid values: 1 to 100.\\
         # Default value: 100.
@@ -2961,6 +2962,7 @@ class DescribeOrgsRequest(TeaModel):
         self.org_name = org_name
         # The parent organization ID.
         self.parent_org_id = parent_org_id
+        self.show_extras = show_extras
 
     def validate(self):
         pass
@@ -2979,6 +2981,8 @@ class DescribeOrgsRequest(TeaModel):
             result['OrgName'] = self.org_name
         if self.parent_org_id is not None:
             result['ParentOrgId'] = self.parent_org_id
+        if self.show_extras is not None:
+            result['ShowExtras'] = self.show_extras
         return result
 
     def from_map(self, m: dict = None):
@@ -2991,6 +2995,64 @@ class DescribeOrgsRequest(TeaModel):
             self.org_name = m.get('OrgName')
         if m.get('ParentOrgId') is not None:
             self.parent_org_id = m.get('ParentOrgId')
+        if m.get('ShowExtras') is not None:
+            self.show_extras = m.get('ShowExtras')
+        return self
+
+
+class DescribeOrgsShrinkRequest(TeaModel):
+    def __init__(
+        self,
+        max_results: int = None,
+        next_token: str = None,
+        org_name: str = None,
+        parent_org_id: str = None,
+        show_extras_shrink: str = None,
+    ):
+        # The maximum number of entries to return. Valid values: 1 to 100.\\
+        # Default value: 100.
+        self.max_results = max_results
+        # The token that determines the start point of the query. The return value is the value of the NextToken response parameter that was returned last time the DescribeOrgs operation was called.
+        self.next_token = next_token
+        # The name of the organization.
+        self.org_name = org_name
+        # The parent organization ID.
+        self.parent_org_id = parent_org_id
+        self.show_extras_shrink = show_extras_shrink
+
+    def validate(self):
+        pass
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        if self.max_results is not None:
+            result['MaxResults'] = self.max_results
+        if self.next_token is not None:
+            result['NextToken'] = self.next_token
+        if self.org_name is not None:
+            result['OrgName'] = self.org_name
+        if self.parent_org_id is not None:
+            result['ParentOrgId'] = self.parent_org_id
+        if self.show_extras_shrink is not None:
+            result['ShowExtras'] = self.show_extras_shrink
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('MaxResults') is not None:
+            self.max_results = m.get('MaxResults')
+        if m.get('NextToken') is not None:
+            self.next_token = m.get('NextToken')
+        if m.get('OrgName') is not None:
+            self.org_name = m.get('OrgName')
+        if m.get('ParentOrgId') is not None:
+            self.parent_org_id = m.get('ParentOrgId')
+        if m.get('ShowExtras') is not None:
+            self.show_extras_shrink = m.get('ShowExtras')
         return self
 
 
@@ -2999,12 +3061,14 @@ class DescribeOrgsResponseBodyOrgs(TeaModel):
         self,
         org_id: str = None,
         org_name: str = None,
+        org_name_path: str = None,
         parent_org_id: str = None,
     ):
         # The organization ID.
         self.org_id = org_id
         # The name of the organizational unit.
         self.org_name = org_name
+        self.org_name_path = org_name_path
         # The parent organization ID.
         self.parent_org_id = parent_org_id
 
@@ -3021,6 +3085,8 @@ class DescribeOrgsResponseBodyOrgs(TeaModel):
             result['OrgId'] = self.org_id
         if self.org_name is not None:
             result['OrgName'] = self.org_name
+        if self.org_name_path is not None:
+            result['OrgNamePath'] = self.org_name_path
         if self.parent_org_id is not None:
             result['ParentOrgId'] = self.parent_org_id
         return result
@@ -3031,6 +3097,8 @@ class DescribeOrgsResponseBodyOrgs(TeaModel):
             self.org_id = m.get('OrgId')
         if m.get('OrgName') is not None:
             self.org_name = m.get('OrgName')
+        if m.get('OrgNamePath') is not None:
+            self.org_name_path = m.get('OrgNamePath')
         if m.get('ParentOrgId') is not None:
             self.parent_org_id = m.get('ParentOrgId')
         return self